requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
langchain>=0.1.0
langgraph>=0.0.20
openai>=1.12.0
duckduckgo-search>=4.1.1
langchain-openai>=0.1.0